refactor wholeprojet
